requirements.txt
numpy==1.17.1
opencv-python==4.1.1.26
Pillow==6.1.0
pycryptodome==3.9.0
